requirements.txt

qrcode 
docxtpl 
python-docx 
fpdf 
reportlab 
pywin32 
numpy
pandas
docx2pdf
streamlit